+++ src/main/java/carracing/model/Cars.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.model;
 
 import java.util.ArrayList;
 import java.util.List;
