+++ src/main/java/carracing/model/MovableStrategy.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.model;
 
 public interface MovableStrategy<T> {
