+++ src/main/java/carracing/model/MoveResult.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.model;
 
 
 public class MoveResult {
