+++ src/main/java/carracing/view/InputInterface.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.view;
 
 public interface InputInterface {
     RacingCondition getInputs();
