+++ src/main/java/carracing/view/OutputInterface.java
@@ -1,4 +1,6 @@
-package carracing;
+package carracing.view;
+
+import carracing.model.MoveResult;
 
 public interface OutputInterface {
     void showCurrentStatus(MoveResult moveResult);
