+++ src/test/java/carracing/view/RacingConditionTest.java
@@ -1,8 +1,10 @@
-package carracing;
+package carracing.view;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import carracing.view.RacingCondition;
 
 public class RacingConditionTest {
